# .forge/evals/agents/fullstack-software-engineer/workspace/iteration-1/eval-task-04-em-reais-contradiz-contrato-centavos/without_skill/run-1/outputs/contracts/openapi/api-recarga.yaml
openapi: 3.1.0
info:
  title: api-recarga
  version: 2.0.0
paths:
  /recargas/{cartaoId}:
    get:
      operationId: listRecargas
      parameters:
        - { name: cartaoId, in: path, required: true, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        "200":
          description: Recargas do cartão, mais recentes primeiro.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Recarga" } }
        "400": { $ref: "#/components/responses/Error" }
  /recargas:
    post:
      operationId: criarRecarga
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NovaRecarga" }
      responses:
        "201":
          description: Recarga criada (assíncrona; nasce em PENDENTE).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recarga" }
        "400": { $ref: "#/components/responses/Error" }
components:
  responses:
    Error:
      description: Erro de validação ou de negócio.
      content:
        application/json:
          schema:
            type: object
            required: [code, message]
            properties: { code: { type: string }, message: { type: string } }
  schemas:
    NovaRecarga:
      type: object
      required: [cartaoId, valor]
      properties:
        cartaoId: { type: string }
        valor: { type: number, format: double, minimum: 0, exclusiveMinimum: true, description: "Valor em reais (BRL), com no máximo duas casas decimais. Ex.: 25.50." }
    Recarga:
      type: object
      required: [id, cartaoId, valor, status, criadaEm]
      properties:
        id: { type: string, format: uuid }
        cartaoId: { type: string }
        valor: { type: number, format: double, description: "Valor em reais (BRL), com no máximo duas casas decimais. Ex.: 25.50." }
        status: { type: string, enum: [PENDENTE, CONFIRMADA, FALHOU] }
        criadaEm: { type: string, format: date-time }